add retry behaviour to redis shipper
The redis shipper is designed to achieve resilience through a proxy
such as twemproxy. In the event of an individual node failure,
twemproxy will propagate the failure to the client, so the client will
need to automatically retry.

However, if all redis nodes are down, the message should be dropped,
to avoid a thundering herd when service is restored. Therefore a retry
limit is put in place.